requirements.txt
pygame-ce
pyserial
pyfirmata2
